orchestrator/requirements.txt
openai>=1.0.0
anthropic>=0.30.0
google-generativeai>=0.5.0
pyyaml>=6.0
